some changes on install scripts

+++ scripts/install.sh
@@ -33,15 +33,21 @@ install_package_if_missing "screen"
 
 # Check and install Java
 hash java 2>/dev/null || {
-    echo >&2 "You need to install Java. Do you want to install it? (y/n)";
+    echo "Do you want to install java? (y/n)"
+
     read wantInstall
     if [ "$wantInstall" = "y" ]; then
+        echo "Select a java version(Minecraft Version 1.7.10 - 1.16.5: Java 8 | 1.9 - 1.16.5: Java 11 | 1.17 - 1.17.1: Java 16 | 1.17 - 1.20.4: Java 17 | 1.20.5 - latest: Java 21)"
+        read javaVersion
         sudo apt install -y software-properties-common
-        sudo echo "deb [arch=amd64] https://some.repository.url focal main" | sudo tee /etc/apt/sources.list.d/adoptium.list > /dev/null
         sudo apt install -y wget apt-transport-https gpg
         sudo wget -qO - https://packages.adoptium.net/artifactory/api/gpg/key/public | sudo gpg --dearmor | sudo tee /etc/apt/trusted.gpg.d/adoptium.gpg > /dev/null
-        sudo echo "deb https://packages.adoptium.net/artifactory/deb $(awk -F= '/^VERSION_CODENAME/{print$2}' /etc/os-release) main" | sudo tee /etc/apt/sources.list.d/adoptium.list
-        sudo apt install -y temurin-22-jdk
+        if [ -f "/etc/os-release" ]; then
+            sudo echo "deb https://packages.adoptium.net/artifactory/deb $(awk -F= '/^VERSION_CODENAME/{print$2}' /etc/os-release) main" | sudo tee /etc/apt/sources.list.d/adoptium.list
+        else
+          sudo echo "deb https://packages.adoptium.net/artifactory/deb $(awk -F= '/^VERSION_CODENAME/{print$2}' /etc/os-release) main" | sudo tee /etc/apt/sources.list.d/adoptium.list
+        fi
+          sudo apt install -y temurin-"$javaVersion"-jdk
     else
         echo "Aborting..."
         exit 1
